CI fixes, also deprecate get_active_tape
