+++ doc/releases/changelog-dev.md
@@ -26,8 +26,12 @@
 
 <h3>Deprecations</h3>
 
-* `qml.tape.stop_recording` and `QuantumTape.stop_recording` are moved to `qml.queuing.QueuingManager.stop_recording`. The old functions will still be available
-  untill v0.29.
+* `qml.tape.stop_recording` and `QuantumTape.stop_recording` are moved to `qml.queuing.QueuingManager.stop_recording`. This
+  method can also be accessed via the alias `qml.queuing.stop_recording`.
+  The old functions will still be available untill v0.29.
+  [(#3068)](https://github.com/PennyLaneAI/pennylane/pull/3068)
+
+* `qml.tape.get_active_tape` is deprecated. Please use `qml.queuing.QueuingContext.active_context()` instead.
   [(#3068)](https://github.com/PennyLaneAI/pennylane/pull/3068)
 
 <h3>Documentation</h3>
