+++ pennylane/transforms/qfunc_transforms.py
@@ -65,7 +65,7 @@ def make_tape(fn):
     """
 
     def wrapper(*args, **kwargs):
-        active_tape = qml.tape.get_active_tape()
+        active_tape = qml.queuing.QueuingContext.active_context()
 
         if active_tape is not None:
             with qml.queuing.stop_recording(), active_tape.__class__() as tape:
