+++ tests/tape/test_tape.py
@@ -1560,6 +1560,8 @@ class TestTapeCopying:
 class TestStopRecording:
     """Test that the stop_recording function works as expected"""
 
+    deprecation_warning = "QuantumTape.stop_recording has moved to qml.queuing.stop_recording"
+
     def test_recording_stopped(self):
         """Test that recording is stopped within a tape context"""
 
@@ -1567,12 +1569,13 @@ class TestStopRecording:
             op0 = qml.RX(0, wires=0)
             assert qml.queuing.QueuingContext.active_context() is tape
 
-            with tape.stop_recording():
-                op1 = qml.RY(1.0, wires=1)
-                assert qml.queuing.QueuingContext.active_context() is None
+            with pytest.warns(UserWarning, self.deprecation_warning):
+                with tape.stop_recording():
+                    op1 = qml.RY(1.0, wires=1)
+                    assert qml.queuing.QueuingContext.active_context() is None
 
-            op2 = qml.RZ(2, wires=1)
-            assert qml.queuing.QueuingContext.active_context() is tape
+                op2 = qml.RZ(2, wires=1)
+                assert qml.queuing.QueuingContext.active_context() is tape
 
         assert len(tape.operations) == 2
         assert tape.operations[0] == op0
@@ -1589,10 +1592,11 @@ class TestStopRecording:
                 assert qml.queuing.QueuingContext.active_context() is tape2
                 op1 = qml.RY(1.0, wires=1)
 
-                with tape2.stop_recording():
-                    assert qml.queuing.QueuingContext.active_context() is None
-                    op2 = qml.RZ(0.6, wires=2)
-                    op3 = qml.CNOT(wires=[0, 2])
+                with pytest.warns(UserWarning, self.deprecation_warning):
+                    with tape2.stop_recording():
+                        assert qml.queuing.QueuingContext.active_context() is None
+                        op2 = qml.RZ(0.6, wires=2)
+                        op3 = qml.CNOT(wires=[0, 2])
 
                 op4 = qml.Hadamard(wires=0)
 
@@ -1615,9 +1619,10 @@ class TestStopRecording:
             op0 = qml.RX(0, wires=0)
             assert qml.queuing.QueuingContext.active_context() is tape
 
-            with tape.stop_recording(), QuantumTape() as temp_tape:
-                assert qml.queuing.QueuingContext.active_context() is temp_tape
-                op1 = qml.RY(1.0, wires=1)
+            with pytest.warns(UserWarning, self.deprecation_warning):
+                with tape.stop_recording(), QuantumTape() as temp_tape:
+                    assert qml.queuing.QueuingContext.active_context() is temp_tape
+                    op1 = qml.RY(1.0, wires=1)
 
             op2 = qml.RZ(2, wires=1)
             assert qml.queuing.QueuingContext.active_context() is tape
@@ -1630,20 +1635,29 @@ class TestStopRecording:
         assert temp_tape.operations[0] == op1
 
 
-def test_gate_tape():
+def test_get_active_tape():
     """Test that the get_active_tape() function returns the currently
     recording tape, or None if no tape is recording"""
-    assert qml.tape.get_active_tape() is None
+    message = (
+        "qml.tape.get_active_tape is now deprecated."
+        " Please use qml.queuing.QueuingContext.active_context"
+    )
+    with pytest.warns(UserWarning, match=message):
+        assert qml.tape.get_active_tape() is None
 
     with QuantumTape() as tape1:
-        assert qml.tape.get_active_tape() is tape1
+        with pytest.warns(UserWarning, match=message):
+            assert qml.tape.get_active_tape() is tape1
 
         with QuantumTape() as tape2:
-            assert qml.tape.get_active_tape() is tape2
+            with pytest.warns(UserWarning, match=message):
+                assert qml.tape.get_active_tape() is tape2
 
-        assert qml.tape.get_active_tape() is tape1
+        with pytest.warns(UserWarning, match=message):
+            assert qml.tape.get_active_tape() is tape1
 
-    assert qml.tape.get_active_tape() is None
+    with pytest.warns(UserWarning, match=message):
+        assert qml.tape.get_active_tape() is None
 
 
 class TestHashing:
